refactor(cli): rename `app` subcommand to `agent`
`agent` is the domain noun of the framework and matches the scaffolded
`agent.py` filename, so `vision-agents agent run` reads as "run the
agent" rather than naming an abstract `app` container. The config
section follows suit:

```toml
[agent]
entrypoint = "agent.py"
```

Renames `cli/app/` → `cli/agent/`, `app_cmd` → `agent_cmd`,
`load_app_config` → `load_agent_config`, the `[app]` TOML section to
`[agent]`, and updates the scaffolded README, init "next steps",
CHANGELOG entry, and tests to match.
