+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## New Features
 
-### `vision-agents` CLI with `init` and `app` commands (#533)
+### `vision-agents` CLI with `init` and `agent` commands (#533)
 
-Adds a console script with two subcommands. `uvx vision-agents init <name>` scaffolds a new agent project (`pyproject.toml`, `agent.py`, `vision-agents.toml`, `.env.example`, `.gitignore`, `README.md`) and runs `uv sync` to provision a venv (skip with `--no-install`). `vision-agents app run|serve` reads `vision-agents.toml` from the project root and forwards args to the project's `Runner.cli()`, so users have a stable entry point even if the agent file is renamed. Templates are rendered with Jinja2.
+Adds a console script with two subcommands. `uvx vision-agents init <name>` scaffolds a new agent project (`pyproject.toml`, `agent.py`, `vision-agents.toml`, `.env.example`, `.gitignore`, `README.md`) and runs `uv sync` to provision a venv (skip with `--no-install`). `vision-agents agent run|serve` reads `vision-agents.toml` from the project root and forwards args to the project's `Runner.cli()`, so users have a stable entry point even if the agent file is renamed. Templates are rendered with Jinja2.
 
 # v0.6.0
