+++ agents-core/vision_agents/cli/agent/__init__.py
@@ -1,0 +1,5 @@
+"""``vision-agents agent`` subcommand."""
+
+from vision_agents.cli.agent.command import agent_cmd
+
+__all__ = ["agent_cmd"]
