+++ agents-core/vision_agents/cli/agent/command.py
@@ -1,4 +1,4 @@
-"""``vision-agents app`` — proxy to the project's Runner CLI."""
+"""``vision-agents agent`` — proxy to the project's Runner CLI."""
 
 import logging
 import subprocess
@@ -7,22 +7,22 @@ from pathlib import Path
 
 import click
 
-from vision_agents.cli.app.config import find_config, load_app_config
+from vision_agents.cli.agent.config import find_config, load_agent_config
 from vision_agents.cli.uv import requires_uv
 
 logger = logging.getLogger(__name__)
 
 
 @click.command(
-    "app",
+    "agent",
     help="Run the project's agent entrypoint (forwards args to its Runner CLI).",
     context_settings={"ignore_unknown_options": True, "allow_extra_args": True},
 )
 @click.argument("args", nargs=-1, type=click.UNPROCESSED)
 @requires_uv
-def app_cmd(args: tuple[str, ...]) -> None:
+def agent_cmd(args: tuple[str, ...]) -> None:
     config_path = find_config(Path.cwd())
-    config = load_app_config(config_path)
+    config = load_agent_config(config_path)
     project_root = config_path.parent
     entrypoint = (project_root / str(config["entrypoint"])).resolve()
     if not entrypoint.is_file():
