--- agents-core/vision_agents/cli/app/__init__.py
@@ -1,5 +1,0 @@
-"""``vision-agents app`` subcommand."""
-
-from vision_agents.cli.app.command import app_cmd
-
-__all__ = ["app_cmd"]
