+++ agents-core/vision_agents/cli/init/templates/README.md.j2
@@ -14,8 +14,8 @@ Generated with `uvx vision-agents init`.
 3. Run the agent:
 
    ```bash
-   uv run vision-agents app run     # single-call console
-   uv run vision-agents app serve   # HTTP server
+   uv run vision-agents agent run     # single-call console
+   uv run vision-agents agent serve   # HTTP server
    ```
 
    Tip: `uv tool install vision-agents` puts the CLI on `PATH` globally, so you can drop the `uv run ` prefix.
