+++ agents-core/vision_agents/cli/init/templates/vision-agents.toml.j2
@@ -1,5 +1,5 @@
 [project]
 name = "{{project_name}}"
 
-[app]
+[agent]
 entrypoint = "agent.py"
